backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
transformers>=4.35.0
torch>=2.1.0
Pillow>=10.1.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
easyocr>=1.7.0
numpy>=1.24.0
opencv-python>=4.8.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
accelerate>=0.25.0
sentencepiece>=0.1.99
num2words>=0.5.13

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

